% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diss.R
\name{diss}
\alias{diss}
\title{dissem.in query}
\usage{
diss(doi = NULL, title = NULL, date = NULL, authors = NULL, ...)
}
\arguments{
\item{doi}{(character) one or more doi's}

\item{title}{(character) title, a work title}

\item{date}{(character) date, a ISO date}

\item{authors}{(character) authors}

\item{...}{curl options passed on to \code{\link[httr]{GET}}
or \code{\link[httr]{POST}}}
}
\description{
dissem.in query
}
\examples{
\dontrun{
diss("10.1016/j.paid.2009.02.013")
diss("10.1186/s40687-015-0044-7")
diss(c("10.1016/j.paid.2009.02.013", "10.1186/s40687-015-0044-7"))

diss(title = "Strange resonances measured in Al+Al collisions at sqrt",
   date = "2015", authors = list(plain = "Lopez, X."))

diss("10.1007/s002200050490")
}
}

