% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diss.R
\name{diss_dois}
\alias{diss_dois}
\title{dissem.in fetch data by DOIs}
\usage{
diss_dois(doi, ...)
}
\arguments{
\item{doi}{(character) one or more DOI's. If you pass more than one DOI,
we do a separate request for each DOI. required.}

\item{...}{curl options passed on to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\description{
dissem.in fetch data by DOIs
}
\examples{
\dontrun{
diss_dois("10.1016/j.paid.2009.02.013")
diss_dois("10.1186/s40687-015-0044-7")
diss_dois(c("10.1016/j.paid.2009.02.013", "10.1186/s40687-015-0044-7"))
}
}
