% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diss_search.R
\name{diss_search}
\alias{diss_search}
\title{dissem.in search}
\usage{
diss_search(query, ...)
}
\arguments{
\item{query}{(character) query terms. required}

\item{...}{curl options passed on to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\description{
dissem.in search
}
\examples{
\dontrun{
diss_search(query = "cellular biology")
}
}
